Separate graeco latin store and palette store
