+++ src/lib/graeco-latin-store.ts
@@ -4,19 +4,11 @@ export type Method = "auto" | "finite" | "cyclic" | "difference" | "direct"
 
 interface GraecoLatinState {
   size: number
-  paletteType: "pastel" | "grayscale" | "scientific_american_59"
-  backgroundShift: number
-  foregroundShift: number
-  paletteSeed: number
   latinMultiplier: number
   greekMultiplier: number
   method: Method
   direct4x4Method: "finite" | "difference"
   setSize: (size: number) => void
-  setPaletteType: (paletteType: "pastel" | "grayscale" | "scientific_american_59") => void
-  setBackgroundShift: (shift: number) => void
-  setForegroundShift: (shift: number) => void
-  setPaletteSeed: (seed: number) => void
   setLatinMultiplier: (multiplier: number) => void
   setGreekMultiplier: (multiplier: number) => void
   setMethod: (method: Method) => void
@@ -25,20 +17,11 @@ interface GraecoLatinState {
 
 export const useGraecoLatinStore = create<GraecoLatinState>((set) => ({
   size: 5,
-  paletteType: "pastel",
-  backgroundShift: 0,
-  foregroundShift: 0,
-  paletteSeed: 0,
   latinMultiplier: 1,
   greekMultiplier: 2,
   method: "auto",
   direct4x4Method: "finite",
   setSize: (size: number) => set({ size }),
-  setPaletteType: (paletteType: "pastel" | "grayscale" | "scientific_american_59") =>
-    set({ paletteType }),
-  setBackgroundShift: (backgroundShift: number) => set({ backgroundShift }),
-  setForegroundShift: (foregroundShift: number) => set({ foregroundShift }),
-  setPaletteSeed: (paletteSeed: number) => set({ paletteSeed }),
   setLatinMultiplier: (latinMultiplier: number) => set({ latinMultiplier }),
   setGreekMultiplier: (greekMultiplier: number) => set({ greekMultiplier }),
   setMethod: (method: Method) => set({ method }),
